Rename signup components
